requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.9
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pillow==10.4.0
openai==1.52.2
itsdangerous==2.2.0
httpx==0.27.2